( chore ) -> API
Separate the API request from the DOMContentLoaded.

Now, each endpoint could have a separete file with differents types of request inside it.

The request may be authenticate or unauthenticated, it will be definie by a parameter in the callApi.

The URI of the endpoint will be always the same, but the endpoint path can change according to the needed